Make Auth Error NSError-ish (#2)
Internally, the Firebase SDK uses some Objective-C → Swift magic with `NS_ERROR_ENUM` to generate a bunch of NSErrors which will get bridges into Swift, so it expects some of those functions to be there as if it were an NSError.

Unfortunately they give this NSError a Swift type that's a little confusing with `NS_SWIFT_NAME(AuthErrorCode)` so our error object is actually called `AuthErrorCode` which we need to match.