Create Auth State Listener API (#4)
This adds the add and remove APIs for auth state listener. Should fix WPP-727.

We've had to allocate our custom subclass on the heap since we want to make it appear as if this type is actually the base `AuthStateListener` instead of our subclass' type. This is important since you will be unable to actually register the listener with the Firebase Auth API unless Swift's interop can figure out that it _is_ an `AuthStateListener`.

Additionally, we've opted to keep the API that requires explicit de-registration of a listener handler, it's a little annoying (since we could also de-register when our opaque type de-initializes) but though it was better to have a 1:1 mapping from the what the underlying SDK expects and what we're providing to the Swift consumer.