FirebaseAuth: improve `Auth.addStateDidChangeListener(_:)`
The reference implementation declares
`AuthStateDidChangeListenerHandler` to be the following:
~~~
public typealias AuthStateDidChangeListenerHandler = NSObjectProtocol
~~~

This most closely resembles `AnyObject`.  Use implicit conversion to
make the handle fully opaque.

Rename the subclassed type identically to the base type as the namespace
disambiguates the types.  This type exists solely to provide a default
method implementation for the pure virtual call of `OnAuthStateChanged`.

As the address of this value is significant, ensure that the type is
marked as a non-copyable, non-assignable, immovable type.  Doing so will
remove the ability of the Swift C++ importer to import the type (which
is why we have the factory method for the type).  Indicate that the type
is an unsafe reference type (we must handle the lifetime of the object
ourselves).

Sink the factory method into the type as `Create`.  Introduce the
associated destructor method `Destroy` to avoid a memory leak.  Due to
the type being a reference type, we cannot easily pass the natively
typed pointer to the method (it will pass by reference).  The
alternative is to provide an instance method to `delete this`.  That is
sometimes done in the dtors, but can be confusing so prefer to use the
`static` method instead.

Finally, as the address is important, create the wrapped callback in the
handle itself to avoid the shuffling of the pointer.  We now create the
callable object inside the handle directly and release it in the `deinit`
of the handle.  This allows us to immediately see the lifetime semantics
of the unsafe pointer.

While tweaking this area, hoist a `guard` statement further to avoid the
casting in the case that we will not use the result anyway.