Auth Errors Should Map To C++ SDK Values Where Possible (#13)
Initially we had used the codes from the Objective-C SDK to power the errors here, however the Objective-C SDK seems to adjust these codes in a lot of cases. We should be using the _names_ of the errors from the Objective-C SDK, but map them to the codes from the auth errors provided from the C++ SDK.

There are some codes from the Objective-C enum that don't map to C++ SDK errors, so I've left them in the `17xxx` number space and left them at the bottom of the enum.