Make CollectionReference implement Query (#32)
Given that C++ types are exposed as structs and structs do not support inheritance in Swift, the fact that the C++ `CollectionReference` inherits from `Query` is not exposed on the Swift side. We then workaround that by defining a protocol that both `Query` and `CollectionReference` can conform to.

Also expose `CollectionReference.path()` and `FieldValue.serverTimestamp()`.