build: add a workaround for Android builds
The android build requires linking to libunwind to prevent underlinking.
This library is squirreled away into the compiler resource directory and
since we do not use the NDK compiler we do not pick it up. Explicitly
add the directory to our linker search path to pick up the necessary
library. This library should really be provided by the NDK as we cannot
ascertain if the unwinder has been modified or contains extensions.